⤵️ Combine models into a common package
