+++ client/src/Components/Sites/Gamemodes/Mixed.tsx
@@ -41,8 +41,6 @@ import tasks from "../../../gamemodes/tasks.json";
 import { TaskUtils } from "../../../helper/TaskUtils";
 import { TruthOrDare } from "../../../gamemodes/TruthOrDare";
 import { ResultPage } from "../../Visuals/ResultPage";
-import { Game } from "../../../helper/models/Game";
-import { Task } from "../../../helper/models/task";
 import { KickList } from "../../Visuals/KickList";
 import { TicUtils } from "../../../gamemodes/tictactoe/TicUtils";
 import { PlayerList } from "../../../helper/models/CustomTypes";
@@ -53,7 +51,9 @@ import { useResult } from "../../../state/actions/resultActions";
 import { useIsHost } from "../../../state/actions/gameActions";
 import { useEvalState, useLeaderboardUpdater, usePollState } from "../../../state/actions/displayStateActions";
 import { EvaluateGame, Serverless } from "../../../helper/Serverless";
-import { Player } from "../../../helper/models/Player";
+import { Game } from "../../../../../common/src/models/Game";
+import { Player } from "../../../../../common/src/models/Player";
+import { Task } from "../../../../../common/src/models/Task";
 
 type TruthOrDareHandle = ElementRef<typeof TruthOrDare>;
 type KickListHandle = ElementRef<typeof KickList>;
