+++ client/src/gamemodes/tictactoe/TicTacToe.tsx
@@ -23,8 +23,8 @@ import { DocumentSnapshot, onSnapshot } from "firebase/firestore";
 import style from "../../css/TicTacToe.module.scss";
 import { TicTacToe as TicTacToeData } from "../../helper/models/TicTacToe";
 import { GameManager } from "../../helper/gameManager";
-import { Player } from "../../helper/models/Player";
 import { firebaseApp } from "../../helper/config";
+import { Player } from "../../../../common/src/models/Player";
 import { TicOptions, TicUtils } from "./TicUtils";
 import { Board } from "./Board";
