+++ client/src/helper/TaskUtils.ts
@@ -15,8 +15,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
+import { IMultiAnswerQuestion, Question } from "../../../common/src/models/Task";
 import { Util } from "./Util";
-import { IMultiAnswerQuestion, Question } from "./models/task";
 
 export class TaskUtils {
     /**
