+++ client/tsconfig.json
@@ -17,5 +17,8 @@
   },
   "include": [
     "./src"
+  ],
+  "references": [
+    { "path": "../common" }
   ]
 }
