+++ common/.gitignore
@@ -1,0 +1,4 @@
+lib/**/*
+node_modules/**/*
+/lib/
+tsconfig.tsbuildinfo
