+++ common/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "common",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "build": "tsc -b"
+  },
+  "author": "Herobone",
+  "license": "GPL-3.0-or-later",
+  "private": true,
+  "dependencies": {
+    "typescript": "^4.4.2"
+  }
+}
