+++ common/src/HostEvents.ts
