+++ common/src/SingleTarget.ts
