+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "functions": "cd functions && npm run build",
     "emulators": "firebase emulators:start",
-    "all": "npm run functions && npm run emulators",
+    "common": "cd common && npm run build",
+    "all": "npm run common && npm run functions && npm run emulators",
     "install-functions": "cd functions && npm i",
     "install-client": "cd client && npm i",
     "install-all": "npm install && npm run install-client && npm run install-functions"
